refactor(backend): rename 3 modules to PascalCase + storage.service (HD-1)
Renamed lowercase module directories to PascalCase for consistency with
the 19 other modules in app/modules/:
- annotations/ → Annotations/
- notes/ → Notes/
- papers/ → Papers/

Also renamed StorageService.ts → storage.service.ts within Papers/
to match camelCase convention (HD-6 consistency).

Updated all 5 import sites: routes/index.ts, Import/import.service.ts,
User/user.service.ts, documentExtractionService.ts, pdfExtractionService.ts,
plus 2 test files.

Phase 8.2.A: Architecture Normalization
