+++ apps/backend/src/app/modules/papers/storage.service.ts
